requirements.txt
appdirs==1.4.4
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
lxml==4.9.2
packaging==23.0
py==1.11.0
pyshark==0.5.3
PyYAML==6.0
requests==2.28.2
urllib3==1.26.14
